mitm-pro-v5/requirements.txt
rich>=13.0
requests>=2.28
pyyaml>=6.0
